chore(main.rs): Added .env file for environment variables
Added a new file `.env` to store environment variables. This file will be used to store configuration settings that are specific to the development environment. The file is intended to be used in conjunction with a configuration management tool, such as dotenv, to ensure that sensitive information is not committed to version control.